Port to FreeBSD and OpenBSD

+++ scripts/run-unittests.sh
@@ -1,29 +1,32 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 main () {
     cd "$(dirname "$(realpath "$0")")/.."
-    local os=$(uname -s)
     if [ -n "$FSARCHS" ]; then
         local archs=()
         IFS=, read -ra archs <<< "$FSARCHS"
         for arch in "${archs[@]}" ; do
             run-tests "$arch"
         done
-    elif [ x$os = xLinux ]; then
-        local cpu=$(uname -m)
-        if [ "x$cpu" == xx86_64 ]; then
-            run-tests linux64
-        elif [ "x$cpu" == xi686 ]; then
-            run-tests linux32
-        else
-            echo "$0: Unknown CPU: $cpu" >&2
-            exit 1
-        fi
-    elif [ "x$os" = xDarwin ]; then
-        run-tests darwin
     else
-        echo "$0: Unknown OS architecture: $os" >&2
-        exit 1
+        local os=$(uname -m -s)
+        case $os in
+            "Darwin arm64")
+                run-tests darwin;;
+            "Darwin x86_64")
+                run-tests darwin;;
+            "FreeBSD amd64")
+                run-tests freebsd_amd64;;
+            "Linux i686")
+                run-tests linux32;;
+            "Linux x86_64")
+                run-tests linux64;;
+            "OpenBSD amd64")
+                run-tests openbsd_amd64;;
+            *)
+                echo "$0: Unknown OS architecture: $os" >&2
+                exit 1
+        esac
     fi
 }
 
@@ -42,14 +45,18 @@ run-tests () {
     echo &&
     echo Start Tests on $arch &&
     echo &&
-    rm -rf stage/$arch/test/log &&
+    rm -rf stage/$arch/workdir &&
+    mkdir stage/$arch/workdir &&
+    if [ "$arch" = openbsd_amd64 ]; then
+        stage/$arch/build/test/fstracetest stage/$arch/workdir
+        return
+    fi
     rm -rf stage/$arch/test/gcov &&
     mkdir -p stage/$arch/test/gcov &&
-    rm -f stage/$arch/test/lock &&
     FSCCFLAGS="-fprofile-arcs -ftest-coverage -O0" \
     FSLINKFLAGS="-fprofile-arcs" \
         ${SCONS:-scons} builddir=test &&
-    stage/$arch/test/test/fstracetest stage/$arch/test &&
+    stage/$arch/test/test/fstracetest stage/$arch/workdir &&
     test-coverage $arch
 }
